Rename testHelpers to index.js
