+++ src/app/components/ErrorPage/index.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { latin, arabic } from '@bbc/gel-foundations/scripts';
-import { shouldMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldMatchSnapshot } from '../../../testHelpers';
 import ErrorMain from './index';
 
 describe('ErrorMain', () => {
