+++ src/app/components/IfAboveIE9Comment/index.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldMatchSnapshot } from '../../../testHelpers';
 import IfAboveIE9 from '.';
 
 describe('Timestamp', () => {
