+++ src/app/components/LinkedData/index.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldShallowMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldShallowMatchSnapshot } from '../../../testHelpers';
 import LinkData from '.';
 
 describe('LinkData', () => {
