+++ src/app/components/Metadata/index.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Metadata from './index';
-import { shouldShallowMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldShallowMatchSnapshot } from '../../../testHelpers';
 
 const metadataSnapshotTest = (
   testDescription,
