+++ src/app/components/ResourceHints/index.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldMatchSnapshot } from '../../../testHelpers';
 import ResourceHints from './index';
 
 describe('ResourceHints', () => {
