+++ src/app/containers/App/index.test.jsx
@@ -3,7 +3,7 @@
  * https://github.com/jtart/react-universal-app
  */
 import React from 'react';
-import { shouldShallowMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldShallowMatchSnapshot } from '../../../testHelpers';
 import { ClientApp, ServerApp } from '.';
 
 describe('ClientApp', () => {
