+++ src/app/containers/Blocks/index.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldShallowMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldShallowMatchSnapshot } from '../../../testHelpers';
 import Blocks from './index';
 import { blockContainingText, singleTextBlock } from '../../models/blocks';
