+++ src/app/containers/Caption/index.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'enzyme';
 import { latin, arabic } from '@bbc/gel-foundations/scripts';
 import CaptionContainer from '.';
-import { shouldMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldMatchSnapshot } from '../../../testHelpers';
 import { ServiceContext } from '../../contexts/ServiceContext';
 import { blockContainingText } from '../../models/blocks';
