+++ src/app/containers/ConsentBanner/Banner/Text.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldMatchSnapshot } from '../../../../testHelpers/testHelpers';
+import { shouldMatchSnapshot } from '../../../../testHelpers';
 import { RequestContextProvider } from '../../../contexts/RequestContext';
 import BannerText from './Text';
