+++ src/app/containers/Header/index.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { news as brandSVG } from '@bbc/psammead-assets/svgs';
 import HeaderContainer from './index';
 import { ServiceContext } from '../../contexts/ServiceContext';
-import { shouldMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldMatchSnapshot } from '../../../testHelpers';
 
 const newsServiceContextStub = {
   brandName: 'BBC News',
