+++ src/app/containers/PageHandlers/withData.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldShallowMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldShallowMatchSnapshot } from '../../../testHelpers';
 import { articleDataNews } from '../Article/fixtureData';
 import WithData from './withData';
