+++ src/app/containers/PageHandlers/withError.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldShallowMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldShallowMatchSnapshot } from '../../../testHelpers';
 import WithError from './withError';
 
 describe('withError HOC', () => {
