+++ src/app/containers/PageHandlers/withLoading.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldShallowMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldShallowMatchSnapshot } from '../../../testHelpers';
 import WithLoading from './withLoading';
 
 describe('withLoading HOC', () => {
