+++ src/app/contexts/DialContext/index.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shouldMatchSnapshot } from '../../../testHelpers/testHelpers';
+import { shouldMatchSnapshot } from '../../../testHelpers';
 import { ComponentUsingContext } from '../../../testHelpers/mockComponents';
 import { DialContext, DialContextProvider } from './index';
